Refactor: Consolidate data sync logic into `SyncDataUseCase`
This commit refactors the data synchronization logic by centralizing it within `SyncDataUseCase` and `ChangePackageUseCase`. Several individual `Update...UseCase` and `Update...ByPackageNameUseCase` classes have been removed, and their logic has been integrated as private functions into these two primary use cases.

This consolidation reduces code duplication, simplifies the dependency graph, and improves the efficiency of data synchronization by fetching data sources (like `launcherAppsActivityInfos` and `appWidgetManagerAppWidgetProviderInfos`) only once per sync operation.

- **`domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/SyncDataUseCase.kt`**:
  - Now contains the logic for a full data sync, previously spread across multiple use cases.
  - New private functions have been added to handle updates for:
    - `EblanApplicationInfos`
    - `EblanAppWidgetProviderInfos`
    - `EblanShortcutInfos`
    - `ApplicationInfoGridItems`
    - `WidgetGridItems`
    - `ShortcutInfoGridItems`
  - Fetches from `launcherAppsWrapper` and `appWidgetManagerWrapper` are now done once at the beginning of the sync process and the results are passed to the update functions.

- **`domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/ChangePackageUseCase.kt`**:
  - Now contains the logic for syncing data for a specific package name, which was previously in separate `...ByPackageNameUseCase` files.
  - New private functions have been added to handle package-specific updates for applications, widgets, and shortcuts.

- **Deleted Use Cases**:
  - Removed numerous redundant use case files that handled specific update operations, as their functionality is now part of `SyncDataUseCase` and `ChangePackageUseCase`. These include:
    - `UpdateApplicationInfoGridItemsUseCase`
    - `UpdateApplicationInfoGridItemsByPackageNameUseCase`
    - `UpdateEblanApplicationInfosUseCase`
    - `UpdateEblanAppWidgetProviderInfosUseCase`
    - `UpdateEblanAppWidgetProviderInfosByPackageNameUseCase`
    - `UpdateEblanShortcutInfosUseCase`
    - `UpdateEblanShortcutInfosByPackageNameUseCase`
    - `UpdateShortcutInfoGridItemsUseCase`
    - `UpdateShortcutInfoGridItemsByPackageNameUseCase`
    - `UpdateWidgetGridItemsUseCase`
    - `UpdateWidgetGridItemsByPackageNameUseCase`
