--- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/UpdateEblanAppWidgetProviderInfosUseCase.kt
@@ -1,131 +1,0 @@
-/*
- *
- *   Copyright 2023 Einstein Blanco
- *
- *   Licensed under the GNU General Public License v3.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *       https://www.gnu.org/licenses/gpl-3.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-package com.eblan.launcher.domain.usecase
-
-import com.eblan.launcher.domain.common.dispatcher.Dispatcher
-import com.eblan.launcher.domain.common.dispatcher.EblanDispatchers
-import com.eblan.launcher.domain.framework.AppWidgetManagerWrapper
-import com.eblan.launcher.domain.framework.FileManager
-import com.eblan.launcher.domain.framework.PackageManagerWrapper
-import com.eblan.launcher.domain.model.EblanAppWidgetProviderInfo
-import com.eblan.launcher.domain.repository.EblanAppWidgetProviderInfoRepository
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.ensureActive
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.withContext
-import java.io.File
-import javax.inject.Inject
-
-class UpdateEblanAppWidgetProviderInfosUseCase @Inject constructor(
-    private val eblanAppWidgetProviderInfoRepository: EblanAppWidgetProviderInfoRepository,
-    private val appWidgetManagerWrapper: AppWidgetManagerWrapper,
-    private val fileManager: FileManager,
-    private val packageManagerWrapper: PackageManagerWrapper,
-    @Dispatcher(EblanDispatchers.IO) private val ioDispatcher: CoroutineDispatcher,
-) {
-    suspend operator fun invoke() {
-        if (!packageManagerWrapper.hasSystemFeatureAppWidgets) return
-
-        withContext(ioDispatcher) {
-            val oldEblanAppWidgetProviderInfos =
-                eblanAppWidgetProviderInfoRepository.eblanAppWidgetProviderInfos.first()
-
-            val newEblanAppWidgetProviderInfos =
-                appWidgetManagerWrapper.getInstalledProviders()
-                    .map { appWidgetManagerAppWidgetProviderInfo ->
-                        ensureActive()
-
-                        val label =
-                            packageManagerWrapper.getApplicationLabel(packageName = appWidgetManagerAppWidgetProviderInfo.packageName)
-
-                        val icon =
-                            packageManagerWrapper.getApplicationIcon(packageName = appWidgetManagerAppWidgetProviderInfo.packageName)
-                                ?.let { byteArray ->
-                                    fileManager.getAndUpdateFilePath(
-                                        directory = fileManager.getFilesDirectory(FileManager.ICONS_DIR),
-                                        name = appWidgetManagerAppWidgetProviderInfo.packageName,
-                                        byteArray = byteArray,
-                                    )
-                                }
-
-                        val preview =
-                            appWidgetManagerAppWidgetProviderInfo.preview?.let { byteArray ->
-                                fileManager.getAndUpdateFilePath(
-                                    directory = fileManager.getFilesDirectory(FileManager.WIDGETS_DIR),
-                                    name = appWidgetManagerAppWidgetProviderInfo.className,
-                                    byteArray = byteArray,
-                                )
-                            }
-
-                        EblanAppWidgetProviderInfo(
-                            className = appWidgetManagerAppWidgetProviderInfo.className,
-                            componentName = appWidgetManagerAppWidgetProviderInfo.componentName,
-                            configure = appWidgetManagerAppWidgetProviderInfo.configure,
-                            packageName = appWidgetManagerAppWidgetProviderInfo.packageName,
-                            targetCellWidth = appWidgetManagerAppWidgetProviderInfo.targetCellWidth,
-                            targetCellHeight = appWidgetManagerAppWidgetProviderInfo.targetCellHeight,
-                            minWidth = appWidgetManagerAppWidgetProviderInfo.minWidth,
-                            minHeight = appWidgetManagerAppWidgetProviderInfo.minHeight,
-                            resizeMode = appWidgetManagerAppWidgetProviderInfo.resizeMode,
-                            minResizeWidth = appWidgetManagerAppWidgetProviderInfo.minResizeWidth,
-                            minResizeHeight = appWidgetManagerAppWidgetProviderInfo.minResizeHeight,
-                            maxResizeWidth = appWidgetManagerAppWidgetProviderInfo.maxResizeWidth,
-                            maxResizeHeight = appWidgetManagerAppWidgetProviderInfo.maxResizeHeight,
-                            preview = preview,
-                            label = label,
-                            icon = icon,
-                        )
-                    }
-
-            if (oldEblanAppWidgetProviderInfos != newEblanAppWidgetProviderInfos) {
-                val eblanAppWidgetProviderInfosToDelete =
-                    oldEblanAppWidgetProviderInfos - newEblanAppWidgetProviderInfos.toSet()
-
-                eblanAppWidgetProviderInfoRepository.upsertEblanAppWidgetProviderInfos(
-                    eblanAppWidgetProviderInfos = newEblanAppWidgetProviderInfos,
-                )
-
-                eblanAppWidgetProviderInfoRepository.deleteEblanAppWidgetProviderInfos(
-                    eblanAppWidgetProviderInfos = eblanAppWidgetProviderInfosToDelete,
-                )
-
-                eblanAppWidgetProviderInfosToDelete.forEach { eblanAppWidgetProviderInfo ->
-                    ensureActive()
-
-                    val icon = File(
-                        fileManager.getFilesDirectory(FileManager.ICONS_DIR),
-                        eblanAppWidgetProviderInfo.packageName,
-                    )
-
-                    val widgetFile = File(
-                        fileManager.getFilesDirectory(FileManager.WIDGETS_DIR),
-                        eblanAppWidgetProviderInfo.className,
-                    )
-
-                    if (icon.exists()) {
-                        icon.delete()
-                    }
-
-                    if (widgetFile.exists()) {
-                        widgetFile.delete()
-                    }
-                }
-            }
-        }
-    }
-}
