--- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/UpdateWidgetGridItemsUseCase.kt
@@ -1,93 +1,0 @@
-/*
- *
- *   Copyright 2023 Einstein Blanco
- *
- *   Licensed under the GNU General Public License v3.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *       https://www.gnu.org/licenses/gpl-3.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-package com.eblan.launcher.domain.usecase
-
-import com.eblan.launcher.domain.common.dispatcher.Dispatcher
-import com.eblan.launcher.domain.common.dispatcher.EblanDispatchers
-import com.eblan.launcher.domain.framework.AppWidgetManagerWrapper
-import com.eblan.launcher.domain.framework.PackageManagerWrapper
-import com.eblan.launcher.domain.model.UpdateWidgetGridItem
-import com.eblan.launcher.domain.model.WidgetGridItem
-import com.eblan.launcher.domain.repository.WidgetGridItemRepository
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.ensureActive
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.withContext
-import javax.inject.Inject
-
-class UpdateWidgetGridItemsUseCase @Inject constructor(
-    private val widgetGridItemRepository: WidgetGridItemRepository,
-    private val appWidgetManagerWrapper: AppWidgetManagerWrapper,
-    private val packageManagerWrapper: PackageManagerWrapper,
-    @Dispatcher(EblanDispatchers.Default) private val defaultDispatcher: CoroutineDispatcher,
-) {
-    suspend operator fun invoke() {
-        if (!packageManagerWrapper.hasSystemFeatureAppWidgets) return
-
-        withContext(defaultDispatcher) {
-            val updateWidgetGridItems = mutableListOf<UpdateWidgetGridItem>()
-
-            val deleteWidgetGridItems = mutableListOf<WidgetGridItem>()
-
-            val widgetGridItems =
-                widgetGridItemRepository.widgetGridItems.first()
-
-            val appWidgetManagerAppWidgetProviderInfos =
-                appWidgetManagerWrapper.getInstalledProviders()
-
-            widgetGridItems.filterNot { widgetGridItem ->
-                widgetGridItem.override
-            }.forEach { widgetGridItem ->
-                ensureActive()
-
-                val appWidgetManagerAppWidgetProviderInfo =
-                    appWidgetManagerAppWidgetProviderInfos.find { appWidgetManagerAppWidgetProviderInfo ->
-                        ensureActive()
-
-                        appWidgetManagerAppWidgetProviderInfo.className == widgetGridItem.className
-                    }
-
-                if (appWidgetManagerAppWidgetProviderInfo != null) {
-                    updateWidgetGridItems.add(
-                        UpdateWidgetGridItem(
-                            id = widgetGridItem.id,
-                            className = appWidgetManagerAppWidgetProviderInfo.className,
-                            componentName = appWidgetManagerAppWidgetProviderInfo.componentName,
-                            configure = appWidgetManagerAppWidgetProviderInfo.configure,
-                            minWidth = appWidgetManagerAppWidgetProviderInfo.minWidth,
-                            minHeight = appWidgetManagerAppWidgetProviderInfo.minHeight,
-                            resizeMode = appWidgetManagerAppWidgetProviderInfo.resizeMode,
-                            minResizeWidth = appWidgetManagerAppWidgetProviderInfo.minResizeWidth,
-                            minResizeHeight = appWidgetManagerAppWidgetProviderInfo.minResizeHeight,
-                            maxResizeWidth = appWidgetManagerAppWidgetProviderInfo.maxResizeWidth,
-                            maxResizeHeight = appWidgetManagerAppWidgetProviderInfo.maxResizeHeight,
-                            targetCellHeight = appWidgetManagerAppWidgetProviderInfo.targetCellWidth,
-                            targetCellWidth = appWidgetManagerAppWidgetProviderInfo.targetCellHeight,
-                        ),
-                    )
-                } else {
-                    deleteWidgetGridItems.add(widgetGridItem)
-                }
-            }
-
-            widgetGridItemRepository.updateWidgetGridItems(updateWidgetGridItems = updateWidgetGridItems)
-
-            widgetGridItemRepository.deleteWidgetGridItemsByPackageName(widgetGridItems = deleteWidgetGridItems)
-        }
-    }
-}
